Added auto centering for location when going to locationDining screen
